+++ src/helpers/index.ts
@@ -3,6 +3,5 @@ export * from './identifiers';
 export * from './interop';
 export * from './io';
 export * from './jsonld';
-export * from './testing';
 export * from './vocabs';
 export * from './wac';