+++ src/main.ts
@@ -1,5 +1,0 @@
-export * from './errors/index';
-export * from './helpers/index';
-export * from './models/index';
-export * from './plugins/index';
-export * from './testing/index';