+++ src/models/SolidDocument.test.ts
@@ -1,5 +1,7 @@
-import { turtleToQuadsSync } from '@/helpers';
-import SolidDocument from '@/models/SolidDocument';
+import { describe, expect, it } from 'vitest';
+
+import SolidDocument from '@noeldemartin/solid-utils/models/SolidDocument';
+import { turtleToQuadsSync } from '@noeldemartin/solid-utils/helpers';
 
 describe('SolidDocument', () => {
 
@@ -13,8 +15,14 @@ describe('SolidDocument', () => {
         /* eslint-disable max-len */
         expect(hasStorageHeader('')).toBe(false);
         expect(hasStorageHeader('<http://www.w3.org/ns/pim/space#Storage>; rel="type"')).toBe(true);
-        expect(hasStorageHeader('<http://www.w3.org/ns/pim/space#Storage>; rel="something-else"; rel="type"')).toBe(true);
-        expect(hasStorageHeader('<http://www.w3.org/ns/pim/space#Storage>; rel="something-else", <http://example.com>; rel="type"')).toBe(false);
+        expect(hasStorageHeader('<http://www.w3.org/ns/pim/space#Storage>; rel="something-else"; rel="type"')).toBe(
+            true,
+        );
+        expect(
+            hasStorageHeader(
+                '<http://www.w3.org/ns/pim/space#Storage>; rel="something-else", <http://example.com>; rel="type"',
+            ),
+        ).toBe(false);
         /* eslint-enable max-len */
     });
 
@@ -27,7 +35,8 @@ describe('SolidDocument', () => {
     it('Parses last modified from document purl:modified', () => {
         const document = new SolidDocument(
             'https://pod.example.org/my-document',
-            turtleToQuadsSync(`
+            turtleToQuadsSync(
+                `
                 <./fallback>
                     <http://purl.org/dc/terms/modified>
                     "2021-09-03T16:23:25.000Z"^^<http://www.w3.org/2001/XMLSchema#dateTime> .
@@ -35,7 +44,9 @@ describe('SolidDocument', () => {
                 <>
                     <http://purl.org/dc/terms/modified>
                     "2021-09-03T16:09:12.000Z"^^<http://www.w3.org/2001/XMLSchema#dateTime> .
-            `, { baseIRI: 'https://pod.example.org/my-document' }),
+            `,
+                { baseIRI: 'https://pod.example.org/my-document' },
+            ),
             new Headers({ 'Last-Modified': 'invalid date' }),
         );
 
@@ -45,7 +56,8 @@ describe('SolidDocument', () => {
     it('Parses last modified from any purl date', () => {
         const document = new SolidDocument(
             'https://pod.example.org/my-document',
-            turtleToQuadsSync(`
+            turtleToQuadsSync(
+                `
                 <./fallback-one>
                     <http://purl.org/dc/terms/modified>
                     "2021-05-03T16:09:12.000Z"^^<http://www.w3.org/2001/XMLSchema#dateTime> .
@@ -53,7 +65,9 @@ describe('SolidDocument', () => {
                 <./fallback-two>
                     <http://purl.org/dc/terms/created>
                     "2021-09-03T16:09:12.000Z"^^<http://www.w3.org/2001/XMLSchema#dateTime> .
-            `, { baseIRI: 'https://pod.example.org/my-document' }),
+            `,
+                { baseIRI: 'https://pod.example.org/my-document' },
+            ),
             new Headers({ 'Last-Modified': 'invalid date' }),
         );