+++ src/models/SolidDocument.ts
@@ -1,7 +1,7 @@
 import { arrayFilter, parseDate, stringMatch } from '@noeldemartin/utils';
-import type { Quad } from 'rdf-js';
+import type { Quad } from '@rdfjs/types';
 
-import { expandIRI } from '@/helpers/vocabs';
+import { expandIRI } from '@noeldemartin/solid-utils/helpers/vocabs';
 
 import SolidStore from './SolidStore';
 
@@ -25,16 +25,13 @@ export default class SolidDocument extends SolidStore {
     }
 
     public isACPResource(): boolean {
-        return !!this.headers.get('Link')
+        return !!this.headers
+            .get('Link')
             ?.match(/<http:\/\/www\.w3\.org\/ns\/solid\/acp#AccessControlResource>;[^,]+rel="type"/);
     }
 
     public isPersonalProfile(): boolean {
-        return !!this.statement(
-            this.url,
-            expandIRI('rdf:type'),
-            expandIRI('foaf:PersonalProfileDocument'),
-        );
+        return !!this.statement(this.url, expandIRI('rdf:type'), expandIRI('foaf:PersonalProfileDocument'));
     }
 
     public isStorage(): boolean {
@@ -54,10 +51,12 @@ export default class SolidDocument extends SolidStore {
     }
 
     public getLastModified(): Date | null {
-        return parseDate(this.headers.get('last-modified'))
-            ?? parseDate(this.statement(this.url, 'purl:modified')?.object.value)
-            ?? this.getLatestDocumentDate()
-            ?? null;
+        return (
+            parseDate(this.headers.get('last-modified')) ??
+            parseDate(this.statement(this.url, 'purl:modified')?.object.value) ??
+            this.getLatestDocumentDate() ??
+            null
+        );
     }
 
     protected expandIRI(iri: string): string {
@@ -65,14 +64,11 @@ export default class SolidDocument extends SolidStore {
     }
 
     private getLatestDocumentDate(): Date | null {
-        const dates = [
-            ...this.statements(undefined, 'purl:modified'),
-            ...this.statements(undefined, 'purl:created'),
-        ]
-            .map(statement => parseDate(statement.object.value))
+        const dates = [...this.statements(undefined, 'purl:modified'), ...this.statements(undefined, 'purl:created')]
+            .map((statement) => parseDate(statement.object.value))
             .filter((date): date is Date => date !== null);
 
-        return dates.length > 0 ? dates.reduce((a, b) => a > b ? a : b) : null;
+        return dates.length > 0 ? dates.reduce((a, b) => (a > b ? a : b)) : null;
     }
 
     private getPermissionsFromWAC(type: string): SolidDocumentPermission[] {