+++ src/plugins/index.ts
@@ -1,2 +1,0 @@
-export * from './chai/index';
-export * from './jest/index';