+++ src/plugins/jest/index.ts
@@ -1,5 +1,0 @@
-import matchers from './matchers';
-
-export function installJestPlugin(): void {
-    expect.extend(matchers);
-}