+++ src/testing/index.ts
@@ -1,2 +1,0 @@
-export * from './mocking';
-export { default as ResponseStub } from './ResponseStub';